pyproject.toml
[project]
name = "midi-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "matplotlib>=3.7.0",
    "numpy>=1.20.0",
    "pretty-midi>=0.2.10",
    "pydantic>=1.8.0",
    "pyyaml>=6.0.2",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "torchvision>=0.15.0",
    "tqdm>=4.60.0",
    "scipy>=1.7.0",
    "rich>=14.1.0",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "python-multipart>=0.0.20",
    "librosa>=0.11.0",
    "setuptools>=80.9.0",
    "mido>=1.3.3",
]
